backend/requirements.txt
Flask==3.1.0
flask-sqlalchemy==3.1.1
flask-cors==4.0.0
tensorflow==2.20.0
keras==3.12.0
numpy==2.1.3
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
Pillow==10.4.0
torch==2.9.1
torchvision==0.24.1
ultralytics==8.3.240
pandas==2.2.3
scikit-learn==1.5.0
seaborn==0.13.2
matplotlib==3.10.0
beautifulsoup4==4.14.3
lxml==5.3.0
requests==2.32.5
gunicorn==21.2.0
pyyaml==6.0.1
gevent==24.10.1